bug solved on mpc_run_solver
Solved a bug that caused errors when using monitors with parametric analyses from python.
The folder for checking if the file monitor exists was changed (line 53), so that line 57 look for the folder "wdir" inside the folder that was already set to "wdir".

Tested both with parametric analyses from python and running single models from STKO.