Change way of computing confined parameters
change way of computing confined parameters for Concrete01, Concrete01WithSITC and Concrete02.
Now the confined version drops to a residual equal to K*fcu (where K is the increasing of peak strength due to confinement). The slope of softening is computed in order to have an energy equivalence respect to Mander model that assumes failure corresponding to failure of the first hoop.